Move utils to utils folder
