+++ src/epomakercontroller/cli.py
@@ -7,7 +7,7 @@ import tkinter as tk
 from .commands.data.constants import Profile
 from .configs.configs import load_main_config
 from .epomakercontroller import EpomakerController
-from .epomaker_utils import print_temp_devices
+from .utils.epomaker_utils import print_temp_devices
 from .keyboard_gui import RGBKeyboardGUI
 
 CONFIG_MAIN = load_main_config()
